Rename Generator->QueueOut and Queue2->Queue.
New CompoundQueue class becomes superclass of RoundRobinQueue.  Instead
of inheriting from QueueOwner, it encapsulates one.  This follows the POLA...
sub-queues have limited access to the CompoundQueue, and the CompoundQueue
doesn't need to expose a public available() method as it would if it inherited
from QueueOwner.

Introduce PriorityQueue, another subclass of CompoundQueue.