Add new Stroke class.  Strokes can (will) be rendered at different LODs, so management of their tesselated mesh is delegated to a StrokeMesh class. This also makes it easy to free memory for pages that aren't currently being viewed: Strokes may simply shed their StrokeMesh(es).  Strokes are responsible for managing the stroke-path, and their material.  StrokeMesh uses a VAO for low-overhead rendering.  Since all strokes share the same topology (currently just a tri-strip), we can share an index-buffer between many strokes, so long as there are enough indices for the longest stroke.
PageView now renders a list of Strokes, no longer using the old Geometry class.

I was investigating a general approach to negotiating the binding between a
mesh and the vertex-attributes required by a Shader.  However, I'm punting on
that for now.  Instead, all stroke-shaders will be subclasses of StrokeShader,
which assumes a fixed stroke vertex-format to bind to.  I commented-out this
work-in-progress... check out Mesh and MeshFormat in mesh.h/cpp.

Refined Shader to make it more convenient to subclass.  Pure virtualfunctions
are provided for subclasses to implement (eg: bindVertexAttribLocations()
is used by StrokeShader to hard-wire the attribute-indices for vertex
position/normal/etc.

Add half-assed Buffer class, with CpuBuffer and GpuBuffer subclasses.
These need work, but will suffice for now.

Fix off-by-one bug in Arc.

Next few commits will be clean-ups.